Moved from double to single precision for all calculations; since the sensors are 16-bit and the AHRS is single-precision, this is possible no loss of actual precision

+++ include/movement.h
@@ -20,7 +20,7 @@ extern "C" {
 #endif
 
 void m_init(void);
-int m_drive(double,unsigned);
+int m_drive(float,float);
 void m_stop(void);
 void m_fini(void);
